pyproject.toml
[project]
name = "doclingextract"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docling>=2.36.1",
    "google-cloud-aiplatform>=1.71.1",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.64",
    "langchain-docling>=1.0.0",
    "langchain-google-vertexai>=2.0.9",
    "langchain-huggingface>=0.3.0",
    "langchain-milvus>=0.2.0",
    "langfuse>=3.0.0",
    "nest-asyncio>=1.6.0",
    "pymilvus>=2.5.11",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.45.1",
    "vertexai>=1.71.1",
]
